+++ packages/terra-status-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update jest test snapshot
+
 ## 4.48.6 - (July 5, 2022)
 
 * Changed